chore: rename `utils/alt` to `utils/flux`
